refactor: move api functions to core/api
