+++ src/core/api/buildPayTransaction.test.ts
@@ -3,12 +3,12 @@ import type { Mock } from 'vitest';
 import {
   CHECKOUT_INVALID_CHARGE_ERROR_MESSAGE,
   UNCAUGHT_CHECKOUT_ERROR_MESSAGE,
-} from '../checkout/constants';
+} from '../../checkout/constants';
 import {
   CDP_CREATE_PRODUCT_CHARGE,
   CDP_HYDRATE_CHARGE,
-} from '../network/definitions/pay';
-import { sendRequest } from '../network/request';
+} from '../../network/definitions/pay';
+import { sendRequest } from '../../network/request';
 /**
  * @vitest-environment node
  */
@@ -28,7 +28,7 @@ import type {
   HydrateChargeAPIParams,
 } from './types';
 
-vi.mock('../network/request');
+vi.mock('../../network/request');
 
 describe('buildPayTransaction', () => {
   afterEach(() => {
