+++ src/core/api/getSwapQuote.ts
@@ -1,28 +1,26 @@
-import { CDP_GET_SWAP_TRADE } from '../network/definitions/swap';
-import { sendRequest } from '../network/request';
-import type { SwapAPIResponse } from '../swap/types';
-import { getSwapErrorCode } from '../swap/utils/getSwapErrorCode';
+import { CDP_GET_SWAP_QUOTE } from '../../network/definitions/swap';
+import { sendRequest } from '../../network/request';
+import type { SwapQuote } from '../../swap/types';
+import { getSwapErrorCode } from '../../swap/utils/getSwapErrorCode';
 import type {
   APIError,
-  BuildSwapTransactionParams,
-  BuildSwapTransactionResponse,
+  GetSwapQuoteParams,
+  GetSwapQuoteResponse,
   SwapAPIParams,
 } from './types';
 import { getAPIParamsForToken } from './utils/getAPIParamsForToken';
-import { getSwapTransaction } from './utils/getSwapTransaction';
 
 /**
- * Retrieves an unsigned transaction for a swap from Token A to Token B.
+ * Retrieves a quote for a swap from Token A to Token B.
  */
-export async function buildSwapTransaction(
-  params: BuildSwapTransactionParams,
-): Promise<BuildSwapTransactionResponse> {
+export async function getSwapQuote(
+  params: GetSwapQuoteParams,
+): Promise<GetSwapQuoteResponse> {
   // Default parameters
   const defaultParams = {
     amountReference: 'from',
     isAmountInDecimals: false,
   };
-
   const apiParamsOrError = getAPIParamsForToken({
     ...defaultParams,
     ...params,
@@ -46,37 +44,27 @@ export async function buildSwapTransaction(
       slippagePercentage = (Number(params.maxSlippage) * 10).toString();
     }
     apiParams = {
-      slippagePercentage,
+      slippagePercentage: slippagePercentage,
       ...apiParams,
     };
   }
 
   try {
-    const res = await sendRequest<SwapAPIParams, SwapAPIResponse>(
-      CDP_GET_SWAP_TRADE,
+    const res = await sendRequest<SwapAPIParams, SwapQuote>(
+      CDP_GET_SWAP_QUOTE,
       [apiParams],
     );
     if (res.error) {
       return {
-        code: getSwapErrorCode('swap', res.error?.code),
+        code: getSwapErrorCode('quote', res.error?.code),
         error: res.error.message,
         message: '',
       };
     }
-
-    const trade = res.result;
-    return {
-      approveTransaction: trade.approveTx
-        ? getSwapTransaction(trade.approveTx, trade.chainId)
-        : undefined,
-      fee: trade.fee,
-      quote: trade.quote,
-      transaction: getSwapTransaction(trade.tx, trade.chainId),
-      warning: trade.quote.warning,
-    };
+    return res.result;
   } catch (_error) {
     return {
-      code: getSwapErrorCode('uncaught-swap'),
+      code: getSwapErrorCode('uncaught-quote'),
       error: 'Something went wrong',
       message: '',
     };
