+++ src/nft/hooks/useMintDetails.ts
@@ -1,6 +1,6 @@
 import { type UseQueryResult, useQuery } from '@tanstack/react-query';
-import { getMintDetails } from '../../api/getMintDetails';
-import type { GetMintDetailsParams, MintDetails } from '../../api/types';
+import { getMintDetails } from '../../core/api/getMintDetails';
+import type { GetMintDetailsParams, MintDetails } from '../../core/api/types';
 import { isNFTError } from '../utils/isNFTError';
 
 export function useMintDetails({
