+++ src/nft/hooks/useTokenDetails.test.ts
@@ -1,10 +1,10 @@
 import { renderHook, waitFor } from '@testing-library/react';
 import { type Mock, beforeEach, describe, expect, it, vi } from 'vitest';
-import { getTokenDetails } from '../../api/getTokenDetails';
+import { getTokenDetails } from '../../core/api/getTokenDetails';
 import { getNewReactQueryTestProvider } from '../../identity/hooks/getNewReactQueryTestProvider';
 import { useTokenDetails } from './useTokenDetails';
 
-vi.mock('../../api/getTokenDetails');
+vi.mock('../../core/api/getTokenDetails');
 
 describe('useTokenDetails', () => {
   beforeEach(() => {
