+++ src/swap/components/SwapProvider.tsx
@@ -9,8 +9,8 @@ import { base } from 'viem/chains';
 import { useAccount, useConfig, useSendTransaction } from 'wagmi';
 import { useSwitchChain } from 'wagmi';
 import { useSendCalls } from 'wagmi/experimental';
-import { buildSwapTransaction } from '../../api/buildSwapTransaction';
-import { getSwapQuote } from '../../api/getSwapQuote';
+import { buildSwapTransaction } from '../../core/api/buildSwapTransaction';
+import { getSwapQuote } from '../../core/api/getSwapQuote';
 import { useCapabilitiesSafe } from '../../internal/hooks/useCapabilitiesSafe';
 import { useValue } from '../../internal/hooks/useValue';
 import { formatTokenAmount } from '../../internal/utils/formatTokenAmount';
